[deoptimizer][wasm] Fix DCHECK for large i31ref constants
This is a follow-up to https://crrev.com/c/5676696 which added the
DCHECK.

FitsInInt32() checks that the 64 bit value fits in an int32_t.
However, what we are interested in for wasm is if the value fits in
a 32 bit value. An i31ref value does not have signedness information,
so it is fine if the highest bit in the 32 bit value gets interpreted
as a sign bit now.

Bug: 42204618
Change-Id: Ia52653073c3a0a3b2e118e382ae6830b77f384a5
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5678418
Commit-Queue: Matthias Liedtke <mliedtke@chromium.org>
Commit-Queue: Darius Mercadier <dmercadier@chromium.org>
Auto-Submit: Matthias Liedtke <mliedtke@chromium.org>
Reviewed-by: Darius Mercadier <dmercadier@chromium.org>
Cr-Commit-Position: refs/heads/main@{#94860}