[riscv] Fix codegen for DIV
RISC-V doesn't raise an exception on divide by zero. The result of dividing by zero is all 1s.
If the divisor register and the destination register are the same, the original value of the register will be overwritten in a division-by-zero operation, making it impossible to determine if the divisor was zero after the division instruction. This CL fix it by adding kUseUniqueRegister attributte to the operand register.

Change-Id: Id2b030f0b731930d611857c1939644ac9c21f3a3
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5681670
Auto-Submit: Yahan Lu <yahan@iscas.ac.cn>
Reviewed-by: Ji Qiu <qiuji@iscas.ac.cn>
Commit-Queue: Ji Qiu <qiuji@iscas.ac.cn>
Cr-Commit-Position: refs/heads/main@{#94902}