[top-level-await] Sync implementation with spec in Module::Evaluate
The implementation and spec text were synced for SourceTextModule but
not Module::Evaluate. The previous sync CL resulted in evaluation of
SourceTextModules that have top-level await or depend on modules with
TLA to transition to the kEvaluatingAsync state. Module::Evaluate was
not updated to understand this state, which was previously conflated
under kEvaluated.

Fixed: 349509539
Change-Id: I5ff02ce1e6f82df9c9e6899a059ddc77308e8821
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5685892
Reviewed-by: Adam Klein <adamk@chromium.org>
Commit-Queue: Shu-yu Guo <syg@chromium.org>
Cr-Commit-Position: refs/heads/main@{#94898}