[heap] Parallelize trivial clearing of maps
This CL allows most weak references pointing to maps to be cleared in
parallel, during the trivial clearing phase. Non-trivial maps are
those contained in Map, TransitionArray or DescriptorArray host
objects.

The triviality check does not come for free, so it is doubtful that
this change improves performance per se. But it is important for
further parallelization in a subsequent CL.

Bug: 348181803
Change-Id: I13776398461b1fd56933d0f554c27d6898e38505
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5660416
Reviewed-by: Omer Katz <omerkatz@chromium.org>
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Nikolaos Papaspyrou <nikolaos@chromium.org>
Cr-Commit-Position: refs/heads/main@{#94922}