[bazel] Enable V8_ENABLE_CONTINUATION_PRESERVED_EMBEDDER_DATA by default
In builds with gn, the flag is enabled by default, even though it can be
disabled in the gn configuration. With this CL, the flag will also be
enabled by default in bazel. However, so far it will not be
configurable. If this is ever needed, it can be introduced later.

R=victorgomes@chromium.org

Change-Id: Id45f716cfdd69e853f61bcf0b3512a140e6531b9
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5686234
Reviewed-by: Victor Gomes <victorgomes@chromium.org>
Auto-Submit: Andreas Haas <ahaas@chromium.org>
Commit-Queue: Andreas Haas <ahaas@chromium.org>
Commit-Queue: Victor Gomes <victorgomes@chromium.org>
Cr-Commit-Position: refs/heads/main@{#94905}