[wasm] Use strong CAS when setting a new fast API call target
So far, `compare_exchange_weak` was used to set up a new API call
target. This is semanically correct, because a failed
`compare_exchange_weak` just means that the fast API optimization does
not get enabled. On PPC, however, the compare-exchange fails once every
6000 runs in a test, which makes the test fail. This compare-exchange is
not performance critical, so we change it to a `compare-exchange-strong`
so that the fast API is always used when possible.

R=thibaudm@chromium.org
CC=mfarazma@redhat.com

Bug: 41492790
Change-Id: Ie07a71133b89070876e0e1e42379a3980d5d9f0c
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5687333
Reviewed-by: Thibaud Michaud <thibaudm@chromium.org>
Commit-Queue: Andreas Haas <ahaas@chromium.org>
Cr-Commit-Position: refs/heads/main@{#94907}