[maglev] Loop SPeeling
Loop SPeeling is speculative loop not-peeling.

This CL adds support for trying to compile loops while speculatively
preserving known_node_aspects and loop phi types. If the speculative
loop header ends up not being compatible with the loop end state, we
treat the current iteration as another peel and compile the loop a
third time -- that time without any optimistic assumptions.

In summary:
1. Normal peel
2. Optimistic loop header, peel if optimism fails
3. Backup actual loop, only needed if 2 fails.

The reason we only apply this strategy to peeled loops is that in a
follow-up CL we will use iteration (1) to track effects to end up with
an optimistic but almost always correct loop header in (2).

Bug: 350764035

Change-Id: I6e60cd5d485374184b7ffbb4878cdc355d2545ff
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5615590
Auto-Submit: Olivier Flückiger <olivf@chromium.org>
Reviewed-by: Victor Gomes <victorgomes@chromium.org>
Commit-Queue: Olivier Flückiger <olivf@chromium.org>
Cr-Commit-Position: refs/heads/main@{#94926}