[sandbox][leaptiering] Create Leaptiering foundation
This CL prepares the foundation for Leaptiering: it creates a new table,
the JSDispatchTable, and adds the necessary infrastructure around it,
in particular basic garbage collection support. The new table is not yet
used in any way. That will happen in follow-up CLs.

The basic idea behind leaptiering is to have a table entry that is
shared between a group of "related" JavaScript functions (essentially
those that can share the same bytecode/code). This entry contains the
current code entrypoint and some more information, for example the
code's expected parameter count (i.e. its signature). This is then used
for two main purposes: (1) to provide fine-grained forward-edge CFI,
both for the V8 Sandbox but also for process-wide CFI and (2) for
seamless tearing without runtime checks: instead of checking on every
function entry whether the current function should tier up or down, with
leaptiering, tiering will simply involve updating the entrypoint in the
table entry. The next call to the function will then automatically use
the new tier. This therefore removes the runtime tiering checks.

Bug: 40931165, 42204201
Change-Id: I450c19d7f6db85a37650b79be73f0be4fca05607
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5672594
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Reviewed-by: Olivier Flückiger <olivf@chromium.org>
Commit-Queue: Samuel Groß <saelo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#94940}