[compiler] Reuse ScopeInfos from previous compiles
- We assign unique IDs to scopes and scope infos mainly based on their
source location, and match newly parsed scopes on existing scope infos
when we allocate scope infos for scopes. Existing scope infos are
found by taking them out from existing inner SFIs from the SFI table
of the script.
- When we streaming compile scripts, we use the
ConstantPoolPointerForwarder to merge in scope infos.
- We add an "eval state" bool to scope infos that we flip for scopes
eval'd in outer scopes. That way we can check the bit to see whether
we're looking at scope infos still belonging to the current script, or
possibly to a parent script that called eval. (In case of nested eval
it's likely possible that the in-between eval either has a context or
doesn't. It doesn't really matter, we look at the actual outer context
on eval and flip that flag. If we need a scope info, that scope info
will have the inverted bit of whatever the outer context had,
independent of its origin).


Change-Id: Ib75b63e46599c8d5fe014275bffb538de21b4ffc
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5630974
Commit-Queue: Toon Verwaest <verwaest@chromium.org>
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Cr-Commit-Position: refs/heads/main@{#94966}