[api] Start deprecation of FastApiTypedArray
There is no advantage of doing the type check and buffer extraction in
generated code, but there is the disadvantage of having to do function
overload resolution dynamically during runtime. Therefore V8 drops the
support for that. The API function can do the type check as efficiently
as V8.

Bug: 41492790
Change-Id: I8ce6718abe66edb324ee38bfa5b94368a5e59a91
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5691777
Commit-Queue: Andreas Haas <ahaas@chromium.org>
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#94971}