[console] Make title parameter handling consistent for various console builtins.
Prior to this CL, the handling of title parameters for various `console`
builtins was highly inconsistent, some would use the empty string as the
default title, some would use the string "default". Also the handling of
`undefined` was different, for some of the builtins, passing `undefined`
would give you a title "undefined", for others it'd be handled as if you
didn't pass any parameters. This was also not only inconsistent within
our own implementation, but also didn't align with what Firefox does.

With this CL we make the behavior consistent in that we always use the
string "default" and we always treat `undefined` the same as a missing
parameter. Conceptually the behavior now is as if it was implemented
with a default parameter like this:

```js
console.time = function(title = "default") {
  // ...
}
```

This CL updates the following builtins:

- `console.time`, `console.timeLog`, `console.timeEnd`
- `console.count`, `console.countReset`,
- `console.timeStamp`,
- `console.profile`, and `console.profileEnd`.

Bug: 350443026
Change-Id: I64ab8cf990b6c81cc7fd07d5ff6324829cdfa458
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5688760
Reviewed-by: Simon Zünd <szuend@chromium.org>
Commit-Queue: Benedikt Meurer <bmeurer@chromium.org>
Auto-Submit: Benedikt Meurer <bmeurer@chromium.org>
Cr-Commit-Position: refs/heads/main@{#94974}