[inspector] Avoid string copies for `consoleTime()` and friends.
This CL changes the interface between the inspector and the embedder to
pass the label parameter for `consoleTime()` and friends as reference
to a `v8::String` instead of a `StringView` in order to avoid another
unnecessary string copy for the common case that tracing is disabled.
This optimization currently only works for `consoleTimeStamp()`, because
for the other three functions we need some additional refactoring to
also avoid copying strings for the purpose of the hash table lookups
and console message generation (that will be done in a follow up CL).

This CL also needs a matching CL on the Chromium side, and a final
cleanup CL on the V8 side to fully migrate to the new interface.

Bug: 350443026, 41433391
Change-Id: I6af8a8c2995ff21931c7823cef35041750883e90
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5701125
Auto-Submit: Benedikt Meurer <bmeurer@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>
Reviewed-by: Simon Zünd <szuend@chromium.org>
Commit-Queue: Benedikt Meurer <bmeurer@chromium.org>
Cr-Commit-Position: refs/heads/main@{#94993}