[wasm] Embed canonical signature id as constant
The canonical signature ID is needed for the type check on indirect
calls. Even though it is a constant, we currently load it from the
WasmTrustedInstanceData in order to make code cachable (because in
another process we might have different canonicalized type IDs).

This CL embeds the constant as a constant in the code, with relocation
information to replace it by a module-local ID on serialization and
replace back by a canonicalized ID on deserialization.

This speeds up indirect calls by ~10% and frees up a field in the
WasmTrustedInstanceData.

R=dlehmann@chromium.org
CC=jkummerow@chromium.org

Bug: 42204513
Change-Id: Ie61fa7e849311566098102f54aed18a47f98846b
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5676329
Reviewed-by: Daniel Lehmann <dlehmann@chromium.org>
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Reviewed-by: Matthias Liedtke <mliedtke@chromium.org>
Cr-Commit-Position: refs/heads/main@{#95051}