[turboshaft][wasm] Fix missing type propagation on br_on_cast_fail fallthrough
For a br_on_cast_fail that always fails, the fallthrough branch is
dynamically unreachable. The type reducer does not figure that
out but can end up in the situation of knowing the source type and
having a static type that doesn't fit to that (i.e. their intersection
is zero), which should not be possible in reachable code.

This fixes the DCHECK by annotating the fallthrough type on
br_on_cast_fail.

There is also a missed optimization as for br_on_cast the analyzer
correctly handles the branch on a condition that is WasmTypeCheckOp
but for br_on_cast_fail the condition is an equality ComparisonOp of a
WasmTypeCheckOp and a Word32Constant(0) which it does not use to
propagate the type check's narrowing.
This is not addressed in this CL.

Fixed: 352720899
Change-Id: I2074dbe84f9cf124ff7e9b5bc28f8bb7e798d491
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5713611
Commit-Queue: Jakob Kummerow <jkummerow@chromium.org>
Auto-Submit: Matthias Liedtke <mliedtke@chromium.org>
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Cr-Commit-Position: refs/heads/main@{#95057}