[platform][cleanup] Remove passed mutex guards
Just a little cleanup before doing the actual change I want to do in the
next CL.
We have lots of "FooLocked()" methods in the code which assume that some
mutex is locked. In the case of the {DefaultForegroundTaskRunner} we
were passing a reference to the {MutexGuard} to those methods to further
enforce that the mutex is held by the caller.

This CL removes this and instead adds a DCHECK that the mutex cannot be
locked again in the caller.

Drive-by: Rename "lock_" to "mutex_".

R=mlippautz@chromium.org

Bug: 42204620
Change-Id: I3738db79d45875c975bd37398ef7e9052dc11896
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5717890
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#95121}