[wasm][fastapi] Add support for word64 parameter
A parameter of a Web API function with the type `long long` in WebIDL
and `int64_t` in C++ still has a type `Number` in its JavaScript
interface.  When the JavaScript interface gets imported into
WebAssembly, the `long long` parameter therefore has to be imported
either as `I32`, `F32`, or `F64`. If it was imported as `I64`, the
WebAPI function would get called with a `BigInt` instead of a `Number`,
which would cause a `TypeError`.

However, neither `I32`, `F32`, nor `F64` match the machine
representation of on `int64_t` in C++. With the basic implementation of
V8's fast API, it would therefore not be possible to call a Web API
function with `long long` parameter, because the machine representation
of the Web API function callback and the type in the signature of the
WebAssembly import definition would never match.

With this CL it is possible to do fast API calls to Web API function
callbacks with `int64_t` and `uint64_t` parameters, even though the
types of the parameters in the signature of the WebAssembly import
definition are `I32`, `F32`, or `F64`. For that, the parameters first
get converted to `int64_t` before they get passed to the API function
callback.

Bug: 41492790
Change-Id: I1fb8b83fb2fdcceacc063c371c09283f3c170390
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5710270
Reviewed-by: Clemens Backes <clemensb@chromium.org>
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Andreas Haas <ahaas@chromium.org>
Cr-Commit-Position: refs/heads/main@{#95102}