[maglev] Revert VO cache in frame states
We currently cache a VO (stored by an interpreter register) of a
deopt frame. The idea was that we didn't need to "search" for the
object inside the frame based on the allocation.

So in deopt points, we swap the InlinedAllocation to the
VirtualObject.

This does not work. The problem is that an allocation points to
the most up-to-date object, which might not be alive in the
current branch.

So the cache is invalidated in following example:

b0:
 Allocate an object VO(0)
 TestBranch(b1, b2)
b1:
 Update object to VO(1)
b2:
 A deopting node.

The inlined allocation in b2 is poiting to VO(1), which is not
alive.

There are potential solutions to this problem, in all of them,
we ended up searching the VO at some point.

For the moment, I am disabling this "optimization" and
implementing the naive approach. We can revise this latter.

This does not change Maglev's compile time on JS2.

Bug: v8:7700
Change-Id: I39801c44ffdbb91311acfd2a6b05cca532e2a6f4
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5720948
Reviewed-by: Darius Mercadier <dmercadier@chromium.org>
Commit-Queue: Victor Gomes <victorgomes@chromium.org>
Auto-Submit: Victor Gomes <victorgomes@chromium.org>
Commit-Queue: Darius Mercadier <dmercadier@chromium.org>
Cr-Commit-Position: refs/heads/main@{#95104}