[sandbox] Add explicit type checks when loading bytecode array from SFI
The SharedFunctionInfo::trusted_data field uses the
kUnknownIndirectPointerTag. As such, any type check on the returned
value must be exhaustive and should not run into an (unchecked) default
case. This CL adds the necessary checks to the routines loading the
BytecodeArray from an SFI.

In the future (with Leaptiering), we may want to load the BytecodeArray
from a function's JSDispatchTable entry, at which point this logic (and
the newly added CHECKs) will no longer be required.

Bug: 40931165
Change-Id: I0d79138de9868a3681ff1e505762aefcc29452f7
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5721169
Commit-Queue: Samuel Groß <saelo@chromium.org>
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Cr-Commit-Position: refs/heads/main@{#95108}