[wasm] Remove experimental prefix from flag name
... for call_indirect inlining. Otherwise, we will later either end up
with a flag having `experimental` in the name that is actually stable,
or will have to rename it and break users.
Neither option is great, so I'd rather not have the prefix and rely on
the implication to `--experimental`. Also, now is still early enough to
change it easily before fuzzing, finching, shipping it.

Bug: 335082212
Change-Id: Iffbad3f1c029fe62faacf9292bbc51dfdb444f9a
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5720949
Reviewed-by: Matthias Liedtke <mliedtke@chromium.org>
Reviewed-by: Clemens Backes <clemensb@chromium.org>
Commit-Queue: Daniel Lehmann <dlehmann@chromium.org>
Cr-Commit-Position: refs/heads/main@{#95113}