[wasm][fastapi] handle out-of-range word64 parameters
When a parameter for a fast API call gets converted from float64 or
float32 to int64 (see https://crrev.com/c/5710270 for an explanation),
the float value may be out of int64 range. In that case, an out-of-range
error should be thrown. This situation was ignored so far. With this CL,
an error gets thrown by doing a regular API call instead of a fast API
call, and let the regular call (actually the regular API call callback)
take care of the range check.

Bug: 41492790
Change-Id: I6c74d7fc97fa0d450fcc202db9ade2f555bec9df
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5712993
Commit-Queue: Andreas Haas <ahaas@chromium.org>
Reviewed-by: Clemens Backes <clemensb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#95115}