[debug] support module syntax in DebugEvaluate::Local
This patch enables module parsing when parsing while debugging
(DebugEvaluate::Evaluate()) when the context the expression is to be
evaluated in occurs within a module source text (determined by walking
the Context chain in search of a Module context).

This will not parse module statement items as part of the expression,
such as import/export specifiers, but does theoretically permit dynamic
imports or import-calls.

There are 2 CCTests added which check that the import.meta is permitted
by DebugEvaluate in a module context, and disallowed in a script
context.

BUG=chromium:40743793

Change-Id: If9834103717cfcbade01fcb9b692fcf761026e5d
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5540688
Reviewed-by: Toon Verwaest <verwaest@chromium.org>
Commit-Queue: Caitlin Potter <caitp@igalia.com>
Cr-Commit-Position: refs/heads/main@{#95171}