Make more runtime functions available to the fuzzers
Previously, we had an explicit allow-list of runtime functions for
fuzzing, and every function that should be available to fuzzers needed
to be added to that. This approach is somewhat fragile, however, as it
is easy to forget adding new functions to that list. With this CL, all
runtime functions that are used for testing are automatically exposed to
fuzzers, but can be manually opted out, which is still necessary for
some functions for a number of reasons.

This CL also makes a number of test functions fuzzer safe, as those are
now exposed to fuzzers. There will likely be a few more functions that
were missed and are not yet fully fuzzer-compatible. However, those
should quickly be identified by our fuzzers, at which point they can
either be made fuzzing-compatible or opted-out of fuzzing.

Bug: 353685107
Change-Id: Ibae038fee2926205bcbc91a8bcadcaec9a8242a1
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5720570
Reviewed-by: Matthias Liedtke <mliedtke@chromium.org>
Reviewed-by: Michael Achenbach <machenbach@chromium.org>
Reviewed-by: Toon Verwaest <verwaest@chromium.org>
Commit-Queue: Samuel Groß <saelo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#95126}