[deoptimizer][wasm][arm64] Re-sign caller pc for pointer authentication
In case of tail calls in the deoptimized frames the caller pc of the
bottom most (outer most) function can end up in a different stack slot
than in the optimized function as the number of parameter stack slots
may differ.

This means, the caller_pc_ needs to be stripped from the PAC and
re-signed again prior to writing it to the output frame.

Fixed: 353582136
Bug: 42204618
Change-Id: I1670c416d3af44af0507b82717ac1df4d3cc11ad
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5716734
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Commit-Queue: Matthias Liedtke <mliedtke@chromium.org>
Reviewed-by: Stephen Röttger <sroettger@google.com>
Cr-Commit-Position: refs/heads/main@{#95148}