Make implicit `this` capture explicit
When declaring a lambda with a value-capture default [=, ...], the
this pointer is implicitly captured by value as well. This results
in potentially-unintuitive behavior and has been deprecated in C++20.
It produces a warning in newer versions of clang
(https://reviews.llvm.org/D142639).

This CL makes the implicit captures explicit, preventing the warning.
It does not change the compiled code at all, since it's just removing
some syntactic sugar.

Bug: 351004963
Change-Id: I5c35d6164aad003e1a9aee1bddcd75e5144ea4b2
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5725522
Reviewed-by: Marja Hölttä <marja@chromium.org>
Commit-Queue: Devon Loehr <dloehr@google.com>
Reviewed-by: Shu-yu Guo <syg@chromium.org>
Cr-Commit-Position: refs/heads/main@{#95323}