Make implicit `this` capture explicit
When declaring a lambda with a value-capture default [=, ...], the
this pointer is implicitly captured by value as well. This results
in potentially-unintuitive behavior and has been deprecated in C++20.
It produces a warning in newer versions of clang
(https://reviews.llvm.org/D142639).

This CL makes the implicit captures explicit, preventing the warning.
It does not change the compiled code at all, since it's just removing
some syntactic sugar.

This CL was uploaded by git cl split.

R=szuend@chromium.org

Bug: 351004963
Change-Id: Ibb1270af7bdc39b2836ce04656a502655446473a
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5726233
Reviewed-by: Thibaud Michaud <thibaudm@chromium.org>
Commit-Queue: Devon Loehr <dloehr@google.com>
Reviewed-by: Simon Zünd <szuend@chromium.org>
Cr-Commit-Position: refs/heads/main@{#95354}