Make implicit `this` capture explicit
When declaring a lambda with a value-capture default [=, ...], the
this pointer is implicitly captured by value as well. This results
in potentially-unintuitive behavior and has been deprecated in C++20.
It produces a warning in newer versions of clang
(https://reviews.llvm.org/D142639).

This CL makes the implicit captures explicit, preventing the warning.
It does not change the compiled code at all, since it's just removing
some syntactic sugar.

This CL was uploaded by git cl split.

R=nicohartmann@chromium.org

Bug: 351004963
Change-Id: I13b7989c85f10e497560648ec0baf4808f5c8a01
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5725496
Commit-Queue: Devon Loehr <dloehr@google.com>
Reviewed-by: Darius Mercadier <dmercadier@chromium.org>
Reviewed-by: Nico Hartmann <nicohartmann@chromium.org>
Cr-Commit-Position: refs/heads/main@{#95191}