Make implicit `this` capture explicit
When declaring a lambda with a value-capture default [=, ...], the
this pointer is implicitly captured by value as well. This results
in potentially-unintuitive behavior and has been deprecated in C++20.
It produces a warning in newer versions of clang
(https://reviews.llvm.org/D142639).

This CL makes the implicit captures explicit, preventing the warning.
It does not change the compiled code at all, since it's just removing
some syntactic sugar.

This CL was uploaded by git cl split.

R=jkummerow@chromium.org

Bug: 351004963
Change-Id: I0da7d1e7fa0c5655df314b73b4e2bcf8bacb67ad
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5725497
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Commit-Queue: Devon Loehr <dloehr@google.com>
Reviewed-by: Toon Verwaest <verwaest@chromium.org>
Cr-Commit-Position: refs/heads/main@{#95164}