Make implicit `this` capture explicit
When declaring a lambda with a value-capture default [=, ...], the
this pointer is implicitly captured by value as well. This results
in potentially-unintuitive behavior and has been deprecated in C++20.
It produces a warning in newer versions of clang
(https://reviews.llvm.org/D142639).

This CL makes the implicit captures explicit, preventing the warning.
It does not change the compiled code at all, since it's just removing
some syntactic sugar.

This CL was uploaded by git cl split.

R=syg@chromium.org

Bug: 351004963
Change-Id: I2c0ea86d2d56ff9ec264ef4f2d7b92b1d5334ba9
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5725498
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Commit-Queue: Devon Loehr <dloehr@google.com>
Reviewed-by: Shu-yu Guo <syg@chromium.org>
Cr-Commit-Position: refs/heads/main@{#95324}