Make implicit `this` capture explicit
When declaring a lambda with a value-capture default [=, ...], the
this pointer is implicitly captured by value as well. This results
in potentially-unintuitive behavior and has been deprecated in C++20.
It produces a warning in newer versions of clang
(https://reviews.llvm.org/D142639).

This CL makes the implicit captures explicit, preventing the warning.
It does not change the compiled code at all, since it's just removing
some syntactic sugar.

This CL was uploaded by git cl split.

R=thibaudm@chromium.org

Bug: 351004963
Change-Id: I9838c2702a634b0fd9935b4a8708335bc9b3c6b4
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5725499
Reviewed-by: Thibaud Michaud <thibaudm@chromium.org>
Commit-Queue: Devon Loehr <dloehr@google.com>
Reviewed-by: Daniel Lehmann <dlehmann@chromium.org>
Cr-Commit-Position: refs/heads/main@{#95356}