[wasm][fastapi] Add test with multiple instances and conflicting imports
Fast API calls from WebAssembly are based on the well-known imports
optimization, which allows faster calls from WebAssembly when the import
is known to be of a special kind. However, this optimization can only
be used when all instances of a WebAssembly module use the same imports.

This CL adds a test where a WebAssembly module gets instantiated with
different fast API call targets to check that the well-known imports
optimization indeed gets disabled.

Bug: 41492790
Change-Id: I91ff611a6cbd26a025dd248cb83e2718b44fd956
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5720910
Reviewed-by: Clemens Backes <clemensb@chromium.org>
Commit-Queue: Andreas Haas <ahaas@chromium.org>
Cr-Commit-Position: refs/heads/main@{#95147}