[wasm] Remove destructor of LogCodesTask
This is a follow-up to https://crrev.com/c/5718451. The destructor is
still sometimes called very late by the platform, after global teardown.
In this case, the WasmEngine was already deallocated.

It turns out we do not need the destructor at all. We can just schedule
a task when the first code to log is put into the queue, similar to how
we request the interrupt.

R=mliedtke@chromium.org

Bug: 42204620
Change-Id: I4daf644bc157f68ac0b53290cc411109c5a6a5c9
Fixed: 354251542
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5729763
Reviewed-by: Matthias Liedtke <mliedtke@chromium.org>
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#95184}