[deoptimizer][wasm] Fix parameter stack slot alignment for v128
While SIMD stack slots are in general not aligned (neither in Liftoff
nor in Turbofan), they are 16-Byte aligned relative to the previous
stack pointer.
The deoptimizer did not account for that.
This is now fixed by calling wasm::IterateSignatureImpl() which should
also ensure that any other implementation details should now be in
sync.

The return stack slots are part of the caller's frame size, so
they should already be accounted for.

Fixed: 353913485
Bug: 42204618
Change-Id: Ie0931690da0ce30bec70e8cfa07643fe128fd622
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5729976
Reviewed-by: Manos Koukoutos <manoskouk@chromium.org>
Commit-Queue: Matthias Liedtke <mliedtke@chromium.org>
Cr-Commit-Position: refs/heads/main@{#95193}