[sandbox] Avoid sbxcheck in JSToWasm and JSToJS Wrapper
The SharedFunctionInfo is currently in a somewhat unfortunate state as
its data field is split into a trusted and an untrusted part. Long term,
we should merge these two fields back together (either by moving all
referenced objects into trusted space, or by separating trusted code
objects from untrusted data objects), but for now this means that we
need to be careful when loading the SFI's data and expecting it to be
trusted. Instead of relying on SBXCHECKs for that purpose, it's better
to explicitly load the trusted part with the expected tag. That way, it
is guaranteed that the returned object is a trusted object of the
expected type, and there is no overhead from a runtime check.

Bug: 354355045
Change-Id: I485ab7d777016105497470f25686c746adaac420
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5729998
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Commit-Queue: Samuel Groß <saelo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#95195}