[turboshaft] Maglev-to-ts: temporarily disable broken test
The test is broken on ia32 because a hole NaN is transformed into a
non-hole NaN when passing it around as a Double.

The proper fix is to use internal::Float64 in Turboshaft's ConstantOp
rather than double, but it will take a bit longer to write, so I'm
landing this small CL to unblock the tree.

Bug: chromium:354893713, chromium:42202729
Change-Id: If95366d28decef0a48de836b12b994b3fd81ecaa
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5735775
Commit-Queue: Darius Mercadier <dmercadier@chromium.org>
Auto-Submit: Darius Mercadier <dmercadier@chromium.org>
Reviewed-by: Matthias Liedtke <mliedtke@chromium.org>
Commit-Queue: Matthias Liedtke <mliedtke@chromium.org>
Cr-Commit-Position: refs/heads/main@{#95213}