[maglev] Fix another bug in extending properties backing stores
This is a follow up to https://chromium-review.googlesource.com/c/v8/v8/+/5680358 .

We don't need a generational write barrier for new -> old references,
but we do need a marking write barrier.

This CL is still keeping the feature behind a flag, since there is a
third bug in it somewhere.

Bug: v8:7700
Change-Id: I8a42731ebc912e8aaadeaaa1c46c07b988674f28
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5729768
Reviewed-by: Victor Gomes <victorgomes@chromium.org>
Commit-Queue: Marja Hölttä <marja@chromium.org>
Cr-Commit-Position: refs/heads/main@{#95235}