[runtime] Fix Object.assign fastcase
The fastcase for `Object.assign({}, src)` relies on the fact that
`{}` is empty, but also does not have any properties in its
prototype chain, that will clash with the copied ones.

To enforce this invariant for cached target maps we add an additional
prototype validity cell to the `Object.assign` fastcase. In other
words to be able to clone `Object.assign(a, b) == c`, the source
b.map must be fast-clonable to c.map, the binary layout of a.map
and c.map must be compatible and the prototype chain of a.map must
be unchanged from the last time we checked.

To accommodate the new validity cell this CL largely rewrites how we
store side-step transitions. Instead of mingling them with normal
transitions on the transition array, we delegate to a dedicated
side step transitions array, which is stored at a fixed offset in
the normal transition array. This removes all the special cases from
the GC.

Bug: 330758172
Fixed: 353886080
Change-Id: If8c2e37169ea8ded8d92146dd318d713d01f96a7
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5695669
Reviewed-by: Toon Verwaest <verwaest@chromium.org>
Commit-Queue: Olivier Flückiger <olivf@chromium.org>
Auto-Submit: Olivier Flückiger <olivf@chromium.org>
Cr-Commit-Position: refs/heads/main@{#95243}