[wasm] Rename WasmApiFunctionRef to WasmImportData
The WasmApiFunctionRef is not only passed for "API functions" (generated
via the JS or C API) any more. For tier-up of import wrappers we now
always generate such an object for every import.

Hence this CL renames them to "WasmImportData". This is passed in place
of the "WasmTrustedInstanceData" as the first parameter to imports.

Since "WasmImportData" was already used for resolved imports, that
existing class is renamed to "ResolvedWasmImport".

R=ahaas@chromium.org, jkummerow@chromium.org
CC=dlehmann@chromium.org

Bug: 354408144
Change-Id: Id255daba8cdc041275a395dd2ab390fc1232efac
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5735183
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Reviewed-by: Matthias Liedtke <mliedtke@chromium.org>
Reviewed-by: Omer Katz <omerkatz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#95279}