Use trusted and untrusted data fields of SFI on all configurations
For the sandbox, the SharedFunctionInfo's data field is split into two
fields: a trusted data field containing a trusted pointer and an
untrusted field containing a tagged pointer. Previously, this split was
only done on sandbox-enabled configurations. Among other things, this
leads to a lot of (more or less) duplicated logic. With this CL, the two
fields now exist on all configurations, which simplifies the code quite
a bit. This also prepares for a follow-up CL that will remove the
(somewhat dangerous) SFI::GetData API in favor of explicit accessors for
the trusted and untrusted data fields.

Bug: 354355045
Change-Id: Ia64303e17e833d4d6c08ef694fdeb54567b9e737
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5741676
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Commit-Queue: Samuel Groß <saelo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#95299}