[debug] Don't CompileTopLevel kEval scripts
When we reuse scope infos for eval scripts, we need to know when we'd
fall off the end of the scope info chain belonging to eval. To do so,
we need to have the topmost SFI (0) of the eval script, and pick up that
SFI's OuterScopeInfo. If we CompileTopLeval an eval script without
SFI(0) present, but any other SFI, we'd accidentally try to inject scope
infos belonging to the parent script.

As a fix: block CompileTopLevel for eval scripts. They won't be
re-executable anyway since the eval cache won't hit.

Change-Id: Ie03bcdd1c8efffae308b65555f0e9d298e6d0d7e
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5741296
Auto-Submit: Toon Verwaest <verwaest@chromium.org>
Commit-Queue: Toon Verwaest <verwaest@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>
Reviewed-by: Simon Zünd <szuend@chromium.org>
Cr-Commit-Position: refs/heads/main@{#95302}