Remove SharedFunctionInfo::GetData
Historically, the SharedFunctionInfo had a single data field, pointing
to some HeapObject. With the introduction of the sandbox, the data field
was split into a trusted_function_data and an untrusted_function_data
field. At that point, the GetData accessor was changed to return
the contents of one of the two fields, depending on which one was
currently in use (an SFI can have either trusted data or untrusted data,
but not both). This is, however, somewhat dangerous as it means that the
object returned by GetData can generally not be trusted as it may come
from the untrusted_function_data field. With this CL, the GetData method
is therefore removed and all users now have to use either GetTrustedData
or GetUntrustedData. Eventually, the two field may again be merged into
a single field (if all data objects move into trusted space). At that
point, the GetUntrustedData method can simply be removed again.

Bug: 354355045
Change-Id: Ib3855fccf0852f320482f6a40a835d6ce35244ba
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5741515
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Commit-Queue: Samuel Groß <saelo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#95332}