[maglev] Track arm/arm64 temps separately from default tmps
The default temporaries (x16,x17) are expected to be available in macro
assembler helpers, and not expected to be needed outside of them. In
particular, the RecordWrite builtin doesn't restore them as part of its
callee-saved calling convension, because they shouldn't have been
allocated in the first place.

We have a concept of temporaries in Maglev, but they are taken out of
the allocatable register set. In this patch, we seperate the temporary
register handling from the "default scratch register" handling, so that
Maglev assembler snippets always operate on a distinct set of
temporaries compared to the macro assembler helpers. This means that we
have to take a bit more care around cloning scratch scopes for deferred
calls, since different assemblers have different ways of maintaining
default scratch registers.

The architecture-independent temporary handling is extracted to a CRTP
base class, and the scope is renamed to a TemporaryRegisterScope, to
match the concept of "temporaries" in the register allocator as distinct
from scratch registers.

This also adds state tracking for the x64 scratch register, to hopefully
reduce the risk of using it twice in a row.

Change-Id: I46022774d7c033e12c23ef507dbfe0c9bee11909
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5739808
Reviewed-by: Olivier Flückiger <olivf@chromium.org>
Commit-Queue: Leszek Swirski <leszeks@chromium.org>
Cr-Commit-Position: refs/heads/main@{#95353}