Make implicit `this` capture explicit
When declaring a lambda with a value-capture default [=, ...], the
this pointer is implicitly captured by value as well. This results
in potentially-unintuitive behavior and has been deprecated in C++20.
It produces a warning in newer versions of clang
(https://reviews.llvm.org/D142639).

This CL makes the implicit captures explicit, preventing the warning.
It does not change the compiled code at all, since it's just removing
some syntactic sugar.

This CL was uploaded by git cl split.

R=leszeks@chromium.org

Bug: 351004963
Change-Id: Ie60bfb3cc59f98b686c302b13f786847adecebd5
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5725523
Reviewed-by: Nico Hartmann <nicohartmann@chromium.org>
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Commit-Queue: Devon Loehr <dloehr@google.com>
Cr-Commit-Position: refs/heads/main@{#95506}