[maglev] Escape receiver in builtin continuations
The builtin continuation call can trigger a stack iteration, which
expects the receiver to be a meterialized object.

Note: for the same reason, we also escape the receiver when
inlining a function, even if the object could be elided.

Fixed: 355493915
Change-Id: Ie6e6e5bb82a1e737cf9beaa370f7a8f00c547a2e
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5748470
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Auto-Submit: Victor Gomes <victorgomes@chromium.org>
Commit-Queue: Leszek Swirski <leszeks@chromium.org>
Cr-Commit-Position: refs/heads/main@{#95398}