[sandbox] Move UncompiledData into trusted space
A SharedFunctionInfo has a trusted_function_data and an
untrusted_function_data field, only one of which can be active. It would
be nice if the two fields could be merged into a single field to
simplify things. For that, all referenced objects would need to be
trusted objects in trusted space. With this CL, UncompiledData objects
are now trusted objects to move towards that goal. In addition, with
UncompiledData now in trusted space, there are no longer any transitions
between trusted and untrusted data after initialization of an SFI. This
furthermore simplifies a few things as the code now no longer needs to
be able to deal with concurrent transitions between these fields.

Bug: 354355045
Change-Id: I911456a5fd5b283ee91bc471c24762fa5edd857f
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5739741
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Samuel Groß <saelo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#95385}