[deoptimizer][wasm][cfi] Always store a signed PC in the FrameDescription
For CFI there may not be any point in time when an attacker can
reasonably replace a PC with a different one without having to sign it.

Prior to this CL the wasm deoptimization mechanism stored the raw PC in
the output FrameDescription first and let it be signed when handling
the next frame. This provides a point in time where an attacker could
(in another thread as during deoptimization no user-code is executed)
find the FrameDescription by going from the Isolate to the Deoptimizer
to the output FrameDescription and replace the PC in a short time
frame.

This CL addresses the issue by only storing already signed PCs in the
FrameDescription. If an attacker replaces the PC, on the next frame the
PointerAuthentication::MoveSignedPC() will fail if the PC does not
authenticate.

Bug: 42204618
Change-Id: I050bfeb58bfa61dfb99d36682f80333859b8da6f
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5749008
Commit-Queue: Matthias Liedtke <mliedtke@chromium.org>
Reviewed-by: Manos Koukoutos <manoskouk@chromium.org>
Reviewed-by: Stephen Röttger <sroettger@google.com>
Cr-Commit-Position: refs/heads/main@{#95406}