[turboshaft] Maglev-to-ts: fix bug with a ZoneRefSet in the wrong zone
The ZoneRefSet input of maglev::CheckMaps was sometimes in the
broker's zone (right after it had been created), which is fine since
this zone is alive until at least ISEL in the TS pipeline, and
sometimes in the MaglevCompilationUnit zone (after it had been copied
in maglev::MergePointInterpreterFrameState::Merge), which is wrong
because this zone gets freed at the end of the Maglev graph building
phase in the Maglev->TS pipeline.

We now always allocate it in the MaglevCompilationUnit zone and copy
it in the Maglev->Turboshaft translation.

Bug: chromium:42204525
Change-Id: Ib7f80cb0df0c7707d49cc1ee8ed6619ee32478f4
Fixed: chromium:356436621
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5748498
Commit-Queue: Leszek Swirski <leszeks@chromium.org>
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Auto-Submit: Darius Mercadier <dmercadier@chromium.org>
Cr-Commit-Position: refs/heads/main@{#95422}