[maglev] Fix integer overflow in keyed access
If we can convert holes to undef, we might have a larger constant
index than allowed by the fixed array.
In this case, we can simply retrun the undefined value.

If we are not allowed to convert holes, it means that the
array is either not a fast access one or we haven't seen a
hole so far according to fbv, which contradicts the fact the
index is a constant larger than the array length.

Fixed: 356441924
Change-Id: Ia54f93306720be6b7ca19829a2ab6703ec37d6b4
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5749094
Auto-Submit: Victor Gomes <victorgomes@chromium.org>
Commit-Queue: Victor Gomes <victorgomes@chromium.org>
Commit-Queue: Olivier Flückiger <olivf@chromium.org>
Reviewed-by: Olivier Flückiger <olivf@chromium.org>
Cr-Commit-Position: refs/heads/main@{#95407}