[wasm][jspi] Remove deprecated JSPI API
The corresponding spec PR is here:
https://github.com/WebAssembly/js-promise-integration/pull/30

The new API is already implemented, so this change only removes the
old one. In particular it removes the WebAssembly.Suspender constructor
and the optional third argument of the WebAssembly.Function constructor.

The suspender object is always implicit in the new API, so a lot of
simplifications follow from this change, especially in the export/import
wrappers.

R=jkummerow@chromium.org

Bug: 42202153
Change-Id: Ia35a7ccf83db83f23adf2d4c46aab2e0a660ecdf
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5755945
Reviewed-by: Omer Katz <omerkatz@chromium.org>
Commit-Queue: Thibaud Michaud <thibaudm@chromium.org>
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Cr-Commit-Position: refs/heads/main@{#95441}