[wasm] Skip NativeModule::EstimateCurrentMemoryConsumption() if unused
EstimateCurrentMemoryConsumption() is a very expensive operation.
If the histogram is disabled (i.e. the result of this expensive
operation is discarded anyway), we shouldn't spend any cycles on it.

As a follow-up it should be reconsidered, whether re-running this on
every single GC mark compact is a good strategy (see TODO).

This fixes the flake for many-modules.js which creates 10k modules,
so for 10k modules we acquire 100'000s of locks on every mark-compact
which seems to prevent the test case from terminating as the main thread
is kept busy executing GCs and doesn't reach an empty task queue when
running in TSAN (which makes locks much more expensive) and the
stress_concurrent_allocation variant.

Fixed: 355462536
Bug: 349610478
Change-Id: I354928bf52369b9abed7f3e18061add70b232029
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5757568
Reviewed-by: Eva Herencsárová <evih@chromium.org>
Commit-Queue: Matthias Liedtke <mliedtke@chromium.org>
Cr-Commit-Position: refs/heads/main@{#95476}