[wasm] Fix uninit call_indirect feedback entry
While the previous check against the first slot of the feedback entry
was correct for call_ref (since initialized entries would never be a
Smi), for call_indirect a genuine call target could happen to coincide
with the uninitialized sentinel zero (with a very low probability).

The fix is to determine whether an entry is uninitialized by the second
slot instead, which is never zero once initialized. This also keeps the
exact same number of comparisons, so performance shouldn't degrade.
Validated locally when forcing the WasmDispatchTable to start at a
address with many low-order zeros.

Drive-by: Very minor change to the feedback processing code to align it
with the feedback updating code in the builtin.

Fixed: 356171295
Change-Id: I66722db2e0a19ffa874d183fe94bcd759b405ff4
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5762884
Commit-Queue: Daniel Lehmann <dlehmann@chromium.org>
Reviewed-by: Matthias Liedtke <mliedtke@chromium.org>
Cr-Commit-Position: refs/heads/main@{#95494}