[sandbox][leaptiering] Allocate dispatch entries for FeedbackCellArray
JSFunctions allocated via a CreateClosure bytecode obtain their
FeedbackCell from a ClosureFeedbackCellArray. With this CL, the
FeedbackCells of such an array now have a valid dispatch entry, and
therefore so do the JSFunctions using them. The dispatch entries are
allocated eagerly, when the ClosureFeedbackCellArray is allocated. This
requires keeping the parameter counts computed during parsing around
until the FeedbackCells are created. While the counts could be stored
elsewhere, too, with this CL they are stored at the end of the
FeedbackMetadata object, after the feedback slot kinds.
The dispatch entries do not yet have a Code object, but that can, in a
future CL, probably just be set to the CompileLazy builtin.

Bug: 40931165, 42204201
Change-Id: Ibf965f84f1259d35655e155102883e426e8e0c3a
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5756889
Commit-Queue: Samuel Groß <saelo@chromium.org>
Reviewed-by: Olivier Flückiger <olivf@chromium.org>
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Cr-Commit-Position: refs/heads/main@{#95523}