[wasm] Always zero-extend incoming i32 register params
We do this everywhere already except in Liftoff on no-sandbox builds.
Make that consistent across all tiers and configurations. This allows
us to remove a redundant unsigned cast in the js-to-wasm wrapper, which
will also make it easier to reuse the torque code for some archs that
expect a sign extension there.

R=ahaas@chromium.org

Change-Id: I58c1afbf16696e3d59fda7dfa00664197059caf6
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5771092
Reviewed-by: Andreas Haas <ahaas@chromium.org>
Commit-Queue: Thibaud Michaud <thibaudm@chromium.org>
Cr-Commit-Position: refs/heads/main@{#95545}