Refactor bootstrapper.cc to pass a functions parameter count
For Leaptiering, every JSFunction needs an entry in the JSDispatchTable.
User-defined functions obtain that entry from their FeedbackCell, but
for builtin functions, the entry needs to be allocated when the
JSFunction is created. However, for that the parameter count must be
known. This CL prepares for that by refactoring the bootstrapping logic
to supply a parameter count (into helper functions such as
InstallFunction) which will then be used in a follow-up CL to allocate
the dispatch table entries. Eventually we should be able to obtain the
parameter count from the builtin Code object, but that is not yet
available. Once it is, the bootstrapping logic can be simplified
further as it then no longer needs to compute the parameter counts.

Bug: 40931165, 42204201, 343498932
Change-Id: If895b5776681a0d6d269b81a76b053e79ad4c9f8
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/5772796
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Commit-Queue: Samuel Groß <saelo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#95548}